Pull CloudInstance out into its own file
Preparation for adding a new version with RootDiskSource.
